Books now take a story from a txt file.
Added a script that splits a text file into the various UI Texts and Dropdowns for the story.  At the moment, the Book does not handle translations.  Translations soon to come.

Note: Be sure to expand the StoryText[] and StoryDrops[] arrays in the scene to accomodate all the string splitting.