added voice over classes and created a StoryDrop prefab.
Added classes to set and play different voice lines.  At the moment, VoiceLines are added onto the dropdown menus and are played when the OnValueChanged event is triggered.  The AudioSource reference is an AS that is added to the Page (ie. add an AudioSource component to the page, and set the VoiceLine reference to it).

-Voice lines are still bugged and need to be fixed.  The wrong voice line is played every time the dropdown is resized (ie. the dropdown's value is changed in-code, so the OnValueChanged event happens).  Need to reprogram how  the dropdowns are resized.

Created a prefab for the dropdowns in the story.  The prefab will have caption images by default that will need to be filled-in, and may have a LayoutElement added to the dropdown as a component.

colored squares are added to the ColorChanger dropdown in TestStory.  It looks much nicer.